pyimagesearch/requirements.txt
black==23.1.0
isort==5.12.0
scikit-image==0.20.0
wandb==0.14.0
pycocotools==2.0.6
timm==0.6.12
fastprogress==1.0.3
torcheval==0.0.6